views: Extract `external_urls` module
