+++ src/lib.rs
@@ -39,6 +39,7 @@ pub mod controllers;
 pub mod db;
 mod downloads_counter;
 pub mod email;
+pub mod external_urls;
 pub mod fastly;
 pub mod github;
 pub mod headers;
