+++ src/views.rs
@@ -1,7 +1,7 @@
 use chrono::NaiveDateTime;
 use secrecy::ExposeSecret;
-use url::Url;
 
+use crate::external_urls::remove_blocked_urls;
 use crate::github;
 use crate::models::{
     ApiToken, Category, Crate, CrateOwnerInvitation, CreatedApiToken, Dependency, DependencyKind,
@@ -12,15 +12,6 @@ use crate::util::rfc3339;
 
 pub mod krate_publish;
 pub use self::krate_publish::{EncodableCrateDependency, PublishMetadata};
-
-/// Hosts in this list are known to not be hosting documentation,
-/// and are possibly of malicious intent e.g. ad tracking networks, etc.
-const DOMAIN_BLOCKLIST: &[&str] = &[
-    "rust-ci.org",
-    "rustless.org",
-    "ironframework.io",
-    "nebulanet.cc",
-];
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct EncodableCategory {
@@ -336,49 +327,6 @@ impl EncodableCrate {
             recent_downloads,
         )
     }
-}
-
-/// Return `None` if the URL host matches a blocked host
-fn remove_blocked_urls(url: Option<String>) -> Option<String> {
-    // Handles if URL is None
-    let url = match url {
-        Some(url) => url,
-        None => return None,
-    };
-
-    // Handles unsuccessful parsing of URL
-    let parsed_url = match Url::parse(&url) {
-        Ok(parsed_url) => parsed_url,
-        Err(_) => return None,
-    };
-
-    // Extract host string from URL
-    let url_host = match parsed_url.host_str() {
-        Some(url_host) => url_host,
-        None => return None,
-    };
-
-    // Match URL host against blocked host array elements
-    if domain_is_blocked(url_host) {
-        None
-    } else {
-        Some(url)
-    }
-}
-
-fn domain_is_blocked(domain: &str) -> bool {
-    DOMAIN_BLOCKLIST
-        .iter()
-        .any(|blocked| &domain == blocked || domain_is_subdomain(domain, blocked))
-}
-
-fn domain_is_subdomain(potential_subdomain: &str, root: &str) -> bool {
-    if !potential_subdomain.ends_with(root) {
-        return false;
-    }
-
-    let root_with_prefix = format!(".{root}");
-    potential_subdomain.ends_with(&root_with_prefix)
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -879,50 +827,5 @@ mod tests {
         assert_some!(json
             .as_str()
             .find(r#""expires_at":"2020-10-24T16:30:00+00:00""#));
-    }
-
-    #[test]
-    fn domain_blocked_no_url_provided() {
-        assert_eq!(remove_blocked_urls(None), None);
-    }
-
-    #[test]
-    fn domain_blocked_invalid_url() {
-        assert_eq!(remove_blocked_urls(Some(String::from("not a url"))), None);
-    }
-
-    #[test]
-    fn domain_blocked_url_contains_partial_match() {
-        assert_eq!(
-            remove_blocked_urls(Some(String::from("http://rust-ci.organists.com")),),
-            Some(String::from("http://rust-ci.organists.com"))
-        );
-    }
-
-    #[test]
-    fn domain_blocked_url() {
-        assert_eq!(
-            remove_blocked_urls(Some(String::from(
-                "http://rust-ci.org/crate/crate-0.1/doc/crate-0.1",
-            ),),),
-            None
-        );
-    }
-
-    #[test]
-    fn domain_blocked_subdomain() {
-        assert_eq!(
-            remove_blocked_urls(Some(String::from(
-                "http://www.rust-ci.org/crate/crate-0.1/doc/crate-0.1",
-            ),),),
-            None
-        );
-    }
-
-    #[test]
-    fn documentation_blocked_non_subdomain() {
-        let input = Some(String::from("http://foorust-ci.org/"));
-        let result = remove_blocked_urls(input);
-        assert_some_eq!(result, "http://foorust-ci.org/");
     }
 }
